backend/requirements.txt
anyio==3.7.1
asgiref==3.8.1
attrs==21.4.0
backoff==1.11.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.5.14
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==4.3.1
django-webpack-loader==3.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
gunicorn==22.0.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.7
lxml==5.2.2
marshmallow==3.21.2
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing-inspect==0.9.0
XlsxWriter==3.2.0
django-environ==0.9.0